+++ src/lib/generate/analysis_pipelines/COPY_NUMBER_ALTERATION/CNVkit.py
@@ -1,5 +1,6 @@
 from support import helper
-from data_type.SEG import segmented_data
+from analysis_pipelines.COPY_NUMBER_ALTERATION.support_functions import fix_chrom, fix_seg_id
+
 
 def main():
     global meta_config
@@ -11,14 +12,12 @@ def main():
     helper.decompress_to_temp(meta_config, study_config, verb)
     helper.working_on(verb)
 
-    helper.working_on(verb, message='Fixing HMMCopy formatting, chromosome, and chromosome max-length ...')
-    segmented_data.fix_hmmcopy_tsv(meta_config, study_config, verb)
-    segmented_data.fix_chrom(meta_config, study_config, verb)
-    segmented_data.fix_hmmcopy_max_chrom(meta_config, study_config, janus_path, verb)
+    helper.working_on(verb, message='Fixing Chromosome numbering ...')
+    fix_chrom(meta_config, study_config, verb)
     helper.working_on(verb)
 
     helper.working_on(verb, message='Fixing .SEG IDs')
-    segmented_data.fix_seg_id(meta_config, study_config, verb)
+    fix_seg_id(meta_config, study_config, verb)
     helper.working_on(verb)
 
     helper.working_on(verb, message='Concating SEG Files to export folder')