+++ src/lib/generate/analysis_pipelines/COPY_NUMBER_ALTERATION/CONTINUOUS_COPY_NUMBER.py
@@ -1,6 +1,5 @@
 from support import helper
-from data_type.DISCRETE_COPY_NUMBER import discrete_copy_number_data
-
+from analysis_pipelines.COPY_NUMBER_ALTERATION.support_functions import discrete_copy_number_data
 
 def main():
     global meta_config