+++ src/lib/generate/analysis_pipelines/MRNA_EXPRESSION/mrna_zscores_data.py
@@ -1,4 +1,4 @@
-"""Remove mRNA Z scores data"""
+"""Support for CAP mRNA Z scores"""
 
 import os