+++ src/lib/generate/analysis_pipelines/MUTATION_EXTENDED/Strelka.py
@@ -1,5 +1,5 @@
 from support import helper
-from data_type.MAF import mutation_data
+from analysis_pipelines.MUTATION_EXTENDED import support_functions
 
 
 def main():
@@ -13,20 +13,16 @@ def main():
     helper.working_on(verb)
 
     helper.working_on('Ensuring both columns exist, otherwise adding UNMATCHED column ...')
-    mutation_data.verify_dual_columns(meta_config, verb)
-    helper.working_on(verb)
-
-    helper.working_on(verb, message='Filtering for only PASS ...')
-    mutation_data.filter_vcf_rejects(meta_config, verb)
+    support_functions.verify_dual_columns(meta_config, verb)
     helper.working_on(verb)
 
     helper.working_on(verb, message='Exporting vcf2maf...')
     helper.working_on(verb, message='And deleting .vcf s...')
-    meta_config = mutation_data.export2maf(meta_config, study_config, verb)
+    meta_config = support_functions.export2maf(meta_config, study_config, verb)
     helper.working_on(verb)
 
     helper.working_on(verb, message='Cleaning MAF Files ...')
-    mutation_data.clean_head(meta_config, verb)
+    support_functions.clean_head(meta_config, verb)
     helper.working_on(verb)
 
     helper.working_on(verb, message='Concating MAF Files to export folder  ...')